test(web): move vitest suites out of tests/ to avoid Python lane false-trigger
The lean workflow's Python test-surface detection matches `tests/**` and
`**/tests/**`. The previous tests/web/*.test.ts (TypeScript) suites matched
that glob, so the gate concluded Python HAD a test surface and failed the
Python coverage lane (no Python coverage config). Relocating the suites to
vitest/ keeps the filename-based jsts detection (**/*.test.*) working while no
longer matching the Python tests/** globs, so the Python lane correctly skips
'no test surface by design'. Imports + vitest.config.ts include glob updated;
coverage remains 100% line+branch.
